OnlineBankingCore: create new base online banking component
